backend/requirements.txt
fastapi[all]
uvicorn[standard]
pydantic
psycopg2-binary
